pyproject.toml
[project]
name = "mandates-core"            # change if this name is taken on PyPI
version = "0.1.2"
description = "Python SDK for creating, signing and verifying Mandates, with remote primitives support."
authors = [{ name = "QuillAI Network" }]
requires-python = ">=3.10"
readme = { file = "README.md", content-type = "text/markdown" }
license = { text = "MIT" }
dependencies = [
  "eth-account>=0.10.0",
  "eth-utils>=2.3.0",
  "requests>=2.31.0",
  "ulid-py>=1.1.0"
]

[project.urls]
Homepage = "https://github.com/quillai-network/mandates-core-python"
Source = "https://github.com/quillai-network/mandates-core-python"
Issues = "https://github.com/quillai-network/mandates-core-python/issues"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["mandates_core"]
